requirements.txt
openai
streamlit
pyttsx3
Pillow
python-dotenv
